pyproject.toml
[project]
name = "retail-banking-rag-v1"
version = "0.1.0"
description = "Retail Banking RAG System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling",
    "fastapi>=0.135.2",
    "langchain",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "python-dotenv>=1.2.2",
    "python-multipart",
    "streamlit",
    "uvicorn>=0.42.0",
]
